Trait actix_net::service::NewService

pub trait NewService where
    <Self::Service as Service>::Request == Self::Request,
    <Self::Service as Service>::Response == Self::Response,
    <Self::Service as Service>::Error == Self::Error,
    <Self::Future as Future>::Item == Self::Service,
    <Self::Future as Future>::Error == Self::InitError, {
    type Request;
    type Response;
    type Error;
    type Service: Service;
    type InitError;
    type Future: Future;
    fn new_service(&self) -> Self::Future;
}

Creates new Service values.

Acts as a service factory. This is useful for cases where new Service values must be produced. One case is a TCP servier listener. The listner accepts new TCP streams, obtains a new Service value using the NewService trait, and uses that new Service value to process inbound requests on that new TCP stream.

Associated Types

type Request

Requests handled by the service

type Response

Responses given by the service

type Error

Errors produced by the service

type Service: Service

The Service value created by this factory

type InitError

Errors produced while building a service.

type Future: Future

The future of the Service instance.

Loading content...
Required methods

fn new_service(&self) -> Self::Future

Create and return a new service value asynchronously.

Loading content...