Trait actix_net::service::ServiceExt

pub trait ServiceExt<Request>: Service<Request> {
    fn apply<T, I, F, Out, Req>(
        self, 
        service: I, 
        f: F
    ) -> AndThen<Self, Apply<T, F, Self::Response, Out, Req>>
    where
        Self: Sized,
        T: Service<Req, Error = Out::Error>,
        I: IntoService<T, Req>,
        F: Fn(Self::Response, &mut T) -> Out,
        Out: IntoFuture<Error = Self::Error>,
    { ... }

    fn and_then<F, B>(self, service: F) -> AndThen<Self, B>
    where
        Self: Sized,
        F: IntoService<B, Self::Response>,
        B: Service<Self::Response, Error = Self::Error>,
    { ... }

    fn from_err<E>(self) -> FromErr<Self, E>
    where
        Self: Sized,
        E: From<Self::Error>,
    { ... }

    fn then<B>(self, service: B) -> Then<Self, B>
    where
        Self: Sized,
        B: Service<Result<Self::Response, Self::Error>, Error = Self::Error>,
    { ... }

    fn map<F, R>(self, f: F) -> Map<Self, F, R>
    where
        Self: Sized,
        F: Fn(Self::Response) -> R,
    { ... }

    fn map_err<F, E>(self, f: F) -> MapErr<Self, F, E>
    where
        Self: Sized,
        F: Fn(Self::Error) -> E,
    { ... }
}

An extension trait for Services that provides a variety of convenient adapters

Provided Methods

fn apply<T, I, F, Out, Req>(
self,
service: I,
f: F
) -> AndThen<Self, Apply<T, F, Self::Response, Out, Req>> where
Self: Sized,
T: Service<Req, Error = Out::Error>,
I: IntoService<T, Req>,
F: Fn(Self::Response, &mut T) -> Out,
Out: IntoFuture<Error = Self::Error>,

Apply function to specified service and use it as a next service in chain.

fn and_then<F, B>(self, service: F) -> AndThen<Self, B> where
Self: Sized,
F: IntoService<B, Self::Response>,
B: Service<Self::Response, Error = Self::Error>,

Call another service after call to this one has resolved successfully.

This function can be used to chain two services together and ensure that the second service isn't called until call to the fist service have finished. Result of the call to the first service is used as an input parameter for the second service's call.

Note that this function consumes the receiving service and returns a wrapped version of it.

fn from_err<E>(self) -> FromErr<Self, E> where
Self: Sized,
E: From<Self::Error>,

fn then<B>(self, service: B) -> Then<Self, B> where
Self: Sized,
B: Service<Result<Self::Response, Self::Error>, Error = Self::Error>,

fn map<F, R>(self, f: F) -> Map<Self, F, R> where
Self: Sized,
F: Fn(Self::Response) -> R,

fn map_err<F, E>(self, f: F) -> MapErr<Self, F, E> where
Self: Sized,
F: Fn(Self::Error) -> E,